refactor: move loader_saver.py to core/data/
